archive/pyproject.toml
[tool.poetry]
name = "coretext"
version = "0.1.0"
description = "CoreText Project"
authors = ["Minh"]
readme = "README.md"
packages = [{include = "coretext"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.0"}
typer = "^0.12.0"
httpx = "^0.27.0"
pydantic = "^2.9.0"
surrealdb = "^1.0.0"
python-multipart = "^0.0.12"
uvicorn = "^0.32.0"
gitpython = "^3.1.43"
sentence-transformers = "^3.2.0"
einops = "^0.8.0"
markdown-it-py = "^4.0.0"
click = "<8.2"
psutil = "^7.2.1"
pyyaml = "^6.0.3"

[tool.poetry.scripts]
coretext = "coretext.cli.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]
